FASTAPI/requirements.txt
fastapi
aioboto3
mangum
uvicorn
python-jose
pydantic-settings
